Fail to translate css grid layout from poster
The old poster was fixed on a 3-column layout due to issues applying
the .col-3 class, etc., from a reveal.js header slide to the parent
section.  This may be a known limitation of pandoc's reveal.js
implementation (or maybe slides?), so consider how to do a workaround.
After all, even if the poster itself isn't reveal.js, changing formats
is just one Rmarkdown flag or command away.