Begin switching template from reveal.js to html_document
The reveal.js template was becoming limiting due to the way that
pandoc handles slides (in a general way), so used the --section-divs
feature of a plain HTML layout instead.  There is still many
references to revealjs in the documentation, CSS, and dependencies,
but the new layout will work better in a general way, and there is
nothing preventing compilation back to the revealjs presentations by
specifying a different format for the output document.