Reference CSS files and other lib_dir resources directly in the template
This change will make it easier to update the template and R function
separately without having to keep track of all of the relevant CSS
files and other resources.  Hopefully it will make the package easier
to maintain due to separating variable template details (pinned
locally to a specific version) and the rmarkdown glue code.  See also
strangeness that results from path backslashes on Windows.