Improve robustness of updating system
Fixed a bug in drposter_update which didn't verify that the resources
in drposter_files/ were successfully moved to a backup.  This resulted
in silently failing updates when a code editor had a file lock on the
CSS files.  Also report detailed version information for drposter to
the cached resource directory for provenance and ease of debugging.