Fix missing fonts in the PDF from FOIT
In the previous commit, some of the Roboto font (non-installed
portion) was not properly embedded into the exported PDF due to a
Flash Of Invisible Text when the page was technically loaded though
the fonts had not finished rendering.  To my understanding, the
virtual time budget flag of 10000 ms gives Chrome up to 10 seconds to
"catch up" and fully finish rendering (which would be especially
helpful if there were complicated Javascript objects, as from
htmlwidgets).

This commit also cleaned up some of the CSS and an old, unused SVG.