Convert html to pdf from R
Adapted my Makefile and the best parts of code from the pagedown, decapitated, and xfun packages to call headless Chrome/Chromium from an R function.  This code was developed on crostini in Chrome OS, so it still needs to be tested in other operating systems.  The RMarkdown hooks and drposter flag also still need to be implemented.