Avoid hanging from Chrome errors
Impose a timeout for the Chrome rendering process (10 seconds by default) so that a Chrome error does not permanently hang the R pipeline as well.  Also more verbosely share the Chrome command and timeout information so that the user knows what to expect (and that the function will not require manual termination).